Apply idms new strategy to EUS and install lanes
EUS: assert ImageDigestMirrorSet has at least the two expected mirror entries
for the upgrade path.

Install: skip applying Konflux IDMS in tests; deployment infra already
provides IDMS entries aligned with the installed version.
