+++ tests/install_upgrade_operators/product_install/conftest.py
@@ -5,23 +5,14 @@ import pytest
 from ocp_resources.cluster_service_version import ClusterServiceVersion
 from ocp_resources.hostpath_provisioner import HostPathProvisioner
 from ocp_resources.hyperconverged import HyperConverged
-from ocp_resources.image_digest_mirror_set import ImageDigestMirrorSet
 from ocp_resources.installplan import InstallPlan
 from ocp_resources.persistent_volume import PersistentVolume
 from ocp_resources.storage_class import StorageClass
-from packaging.version import Version
 from pytest_testconfig import py_config
 from timeout_sampler import TimeoutSampler
 
-from tests.install_upgrade_operators.constants import (
-    KONFLUX_IDMS_NAME,
-    KONFLUX_MIRROR_BASE_URL,
-)
 from tests.install_upgrade_operators.product_install.constants import (
     HCO_NOT_INSTALLED_ALERT,
-)
-from tests.install_upgrade_operators.utils import (
-    apply_konflux_idms,
 )
 from utilities.constants import (
     CRITICAL_STR,
@@ -62,32 +53,6 @@ from utilities.storage import (
 INSTALLATION_VERSION_MISMATCH = "98"
 LOCAL_BLOCK_HPP = "local-block-hpp"
 LOGGER = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="module")
-def installed_konflux_idms(
-    admin_client,
-    is_production_source,
-    cnv_version_to_install_info,
-    nodes,
-    machine_config_pools,
-    machine_config_pools_conditions_scope_module,
-):
-    if is_production_source:
-        LOGGER.info("Production source install, IDMS update not needed.")
-        return
-
-    version = Version(version=cnv_version_to_install_info["version"])
-    required_mirrors = [f"{KONFLUX_MIRROR_BASE_URL}/v{version.major}-{version.minor}"]
-
-    idms = ImageDigestMirrorSet(name=KONFLUX_IDMS_NAME, client=admin_client)
-    apply_konflux_idms(
-        idms=idms,
-        required_mirrors=required_mirrors,
-        machine_config_pools=machine_config_pools,
-        mcp_conditions=machine_config_pools_conditions_scope_module,
-        nodes=nodes,
-    )
 
 
 @pytest.fixture(scope="module")
@@ -181,7 +146,6 @@ def cnv_install_plan_installed(
 def installed_openshift_virtualization(
     admin_client,
     disabled_default_sources_in_operatorhub_scope_module,
-    installed_konflux_idms,
     hyperconverged_catalog_source,
     created_cnv_namespace,
     created_cnv_operator_group,
